# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZeroQuant Trading API",
    "description": "\n# ZeroQuant 트레이딩 봇 REST API\n\n전략 관리, 백테스트, 포트폴리오 분석을 위한 REST API입니다.\n\n## 주요 기능\n\n- **전략 관리**: 트레이딩 전략 생성, 조회, 시작/중지\n- **백테스트**: 과거 데이터 기반 전략 성과 분석\n- **포트폴리오**: 실시간 포트폴리오 상태 조회\n- **시장 데이터**: 실시간 시세 및 차트 데이터\n- **모니터링**: 에러 추적 및 시스템 상태 모니터링\n- **스크리닝**: 종목 스크리닝 및 필터링\n- **ML**: 머신러닝 모델 훈련 및 배포\n\n## 인증\n\n대부분의 엔드포인트는 JWT Bearer 토큰 인증이 필요합니다.\n`Authorization: Bearer <token>` 헤더를 포함하세요.\n\n## 심볼 동기화\n\n- **KRX**: `POST /api/v1/dataset/sync/krx` - 한국 거래소 심볼\n- **EODData**: `POST /api/v1/dataset/sync/eod` - 해외 거래소 심볼\n",
    "contact": {
      "name": "ZeroQuant Team",
      "url": "https://github.com/user/zeroquant"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.4.4"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "로컬 개발 서버"
    }
  ],
  "paths": {
...
    "/api/v1/ranking/7factor/batch": {
      "post": {
        "tags": [
          "ranking"
        ],
        "summary": "POST /api/v1/ranking/7factor/batch - 7Factor 일괄 조회",
        "operationId": "get_seven_factor_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SevenFactorBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "7Factor 일괄 데이터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SevenFactorBatchResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
    "/api/v1/ranking/7factor/{ticker}": {
      "get": {
        "tags": [
          "ranking"
        ],
        "summary": "GET /api/v1/ranking/7factor/{ticker} - 특정 종목의 7Factor 데이터 조회",
        "operationId": "get_seven_factor",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "description": "종목 티커",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "market",
            "in": "query",
            "description": "시장 (기본: KR)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "7Factor 데이터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SevenFactorResponse"
                }
              }
            }
          },
          "404": {
            "description": "종목 없음"
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
    "/api/v1/ranking/global": {
      "post": {
        "tags": [
          "ranking"
        ],
        "summary": "POST /api/v1/ranking/global - 모든 심볼 GlobalScore 계산",
        "operationId": "calculate_global",
        "responses": {
          "200": {
            "description": "계산 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculateResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
    "/api/v1/ranking/top": {
      "get": {
        "tags": [
          "ranking"
        ],
        "summary": "GET /api/v1/ranking/top - 상위 랭킹 조회",
        "operationId": "get_top_ranked",
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "description": "시장 필터 (KR, US 등)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "grade",
            "in": "query",
            "description": "등급 필터 (BUY, WATCH 등)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "description": "최소 점수",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "반환 개수 (기본 50, 최대 500)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "route_state",
            "in": "query",
            "description": "RouteState 필터 (ATTACK, ARMED, WATCH, REST)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "랭킹 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalculateResponse": {
        "type": "object",
        "description": "GlobalScore 계산 응답",
        "required": [
          "processed",
          "started_at",
          "completed_at"
        ],
        "properties": {
          "completed_at": {
            "type": "string",
            "description": "계산 완료 시간"
          },
          "processed": {
            "type": "integer",
            "format": "int32",
            "description": "처리된 종목 수"
          },
          "started_at": {
            "type": "string",
            "description": "계산 시작 시간"
          }
        }
      },
...
      "FilterInfo": {
        "type": "object",
        "description": "필터 정보",
        "required": [
          "limit"
        ],
        "properties": {
          "grade": {
            "type": [
              "string",
              "null"
            ]
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "market": {
            "type": [
              "string",
              "null"
            ]
          },
          "min_score": {
            "type": [
              "string",
              "null"
            ]
          },
          "route_state": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RankedSymbol": {
        "type": "object",
        "description": "GlobalScore 랭킹 응답용 (JOIN with symbol_info)",
        "required": [
          "ticker",
          "name",
          "market",
          "overall_score",
          "grade",
          "component_scores",
          "calculated_at"
        ],
        "properties": {
          "calculated_at": {
            "type": "string",
            "format": "date-time"
          },
          "component_scores": {
            "$ref": "#/components/schemas/Value"
          },
          "confidence": {
            "type": [
              "string",
              "null"
            ]
          },
          "exchange": {
            "type": [
              "string",
              "null"
            ]
          },
          "grade": {
            "type": "string"
          },
          "market": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "overall_score": {
            "type": "string",
            "description": "종합 점수 (0-100) - JSON에서 숫자로 직렬화"
          },
          "penalties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Value"
              }
            ]
          },
          "route_state": {
            "type": [
              "string",
              "null"
            ],
            "description": "RouteState (실시간 계산됨, DB 조회 시 None)"
          },
          "ticker": {
            "type": "string"
          }
        }
      },
...
      "RankingResponse": {
        "type": "object",
        "description": "랭킹 조회 응답",
        "required": [
          "symbols",
          "total",
          "filter"
        ],
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/FilterInfo",
            "description": "필터 정보"
          },
          "symbols": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RankedSymbol"
            },
            "description": "종목 목록"
          },
          "total": {
            "type": "integer",
            "description": "총 개수",
            "minimum": 0
          }
        }
      },
...
      "SevenFactorBatchRequest": {
        "type": "object",
        "description": "7Factor 일괄 조회 요청",
        "required": [
          "tickers"
        ],
        "properties": {
          "market": {
            "type": "string",
            "description": "시장 (기본: KR)"
          },
          "tickers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "티커 목록"
          }
        }
      },
      "SevenFactorBatchResponse": {
        "type": "object",
        "description": "7Factor 일괄 조회 응답",
        "required": [
          "factors",
          "total"
        ],
        "properties": {
          "factors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SevenFactorResponse"
            },
            "description": "7Factor 데이터 목록"
          },
          "total": {
            "type": "integer",
            "description": "총 개수",
            "minimum": 0
          }
        }
      },
      "SevenFactorData": {
        "type": "object",
        "description": "7Factor 데이터",
        "required": [
          "norm_momentum",
          "norm_value",
          "norm_quality",
          "norm_volatility",
          "norm_liquidity",
          "norm_growth",
          "norm_sentiment"
        ],
        "properties": {
          "norm_growth": {
            "type": "string",
            "description": "성장성 (0-100)"
          },
          "norm_liquidity": {
            "type": "string",
            "description": "유동성 (0-100)"
          },
          "norm_momentum": {
            "type": "string",
            "description": "모멘텀 (0-100)"
          },
          "norm_quality": {
            "type": "string",
            "description": "품질 (0-100)"
          },
          "norm_sentiment": {
            "type": "string",
            "description": "시장 심리 (0-100)"
          },
          "norm_value": {
            "type": "string",
            "description": "가치 (0-100)"
          },
          "norm_volatility": {
            "type": "string",
            "description": "변동성 (0-100, 낮은 변동성 = 높은 점수)"
          }
        }
      },
...
      "SevenFactorResponse": {
        "type": "object",
        "description": "7Factor 응답용 타입",
        "required": [
          "ticker",
          "name",
          "market",
          "factors",
          "composite_score",
          "calculated_at"
        ],
        "properties": {
          "calculated_at": {
            "type": "string",
            "format": "date-time",
            "description": "계산 시각"
          },
          "composite_score": {
            "type": "string",
            "description": "종합 점수 (0-100) - JSON에서 숫자로 직렬화"
          },
          "factors": {
            "$ref": "#/components/schemas/SevenFactorData",
            "description": "7개 정규화 팩터 (0-100)"
          },
          "global_score": {
            "type": [
              "string",
              "null"
            ],
            "description": "기존 GlobalScore 정보 - JSON에서 숫자로 직렬화"
          },
          "grade": {
            "type": [
              "string",
              "null"
            ]
          },
          "market": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ticker": {
            "type": "string"
          }
        }
      },
...
      "Value": {}
    }
  },
...
}